+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `filament-notification-sound` will be documented in this file.
 
-## 1.0.0 - 202X-XX-XX
+## 1.0.0 - 2025-11-30
 
-- initial release
+- Initial release
+- Added sound notifications for Filament database notifications
+- Configurable sound path, volume, and animation